// src/app/interfaces/conduccion-interface.ts
// Generated by https://quicktype.io

export interface ResultConducciones {
    ok:         boolean;
    msg:        string;
    conduccion: Conduccion[];
}

export interface ResultConduccion {
    ok:         boolean;
    msg:        string;
    conduccion: Conduccion;
}

export interface Conduccion {
    _id:      string;
    chofer:   Chofer;
    vehiculo: Vehiculos;
}
export interface ResultConduccionPer {
    ok:         boolean;
    msg:        string;
    conduccion: ConduccionPer;
}
export interface ConduccionPer {
    _id:      string;
    chofer:   string;
    vehiculo: string;
}
export interface Chofer {
    _id:    string;
    nombre: string;
}

export interface Vehiculos {
    _id:     string;
    nombre:  string;
    marca:   string;
    placa:   string;
    color:   string;
    ano:     string;
    estado:  boolean;
    usuario: string;
    __v:     number;
}